feat(UserFactory.java, Roles.java): add UserFactory class and Roles enum to create User objects with different roles
The UserFactory class was added to create User objects with different roles. The Roles enum was also added to define the different roles that a User can have. The createUser method in the UserFactory class takes in the name, email, password, and role of the User and returns a User object with the specified role. The createTeacher and createSecretary methods in the UserFactory class create Teacher and Secretary objects respectively with the specified name, email, and password.