feat(tests): add MongoDB configuration for integration tests
Add a new configuration class `MongoConfig` to configure MongoDB for integration tests. The class sets up a `MongoTemplate` and a `MongodExecutable` bean. The `MongodExecutable` bean is used to start and stop an embedded MongoDB instance.

Add two new annotations `@ConfigContainersTest` and `@ConfigMongoDBTest` to the test classes. The `@ConfigContainersTest` annotation is used to activate the `test` profile. The `@ConfigMongoDBTest` annotation is used to import the `MongoConfig` class.